pythonScripts/requirements.txt
contourpy==1.3.2
cycler==0.12.1
dt-apriltags==3.1.7
exif==1.6.1
fonttools==4.60.1
ImageIO==2.37.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plum-py==0.8.7
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
rawpy==0.25.1
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
